+++ packages/cloud_firestore/lib/src/firestore.dart
@@ -118,4 +118,12 @@ class Firestore {
     });
     return result?.cast<String, dynamic>() ?? <String, dynamic>{};
   }
+
+  Future<void> enablePersistence(bool enable) async {
+    assert(enable != null);
+    await channel.invokeMethod('Firestore#enablePersistence', <String, dynamic>{
+      'app': app.name,
+      'enable': enable,
+    });
+  }
 }