+++ packages/cloud_functions/example/test/widget_test.dart
@@ -7,7 +7,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:cloud_functions_example/main.dart';
+import '../lib/main.dart';
 
 void main() {
   testWidgets('Verify Response', (WidgetTester tester) async {